+++ src/main/java/org/mmarini/wheelly/apps/BatchTraining.java
@@ -27,8 +27,6 @@ package org.mmarini.wheelly.apps;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.reactivex.rxjava3.core.Completable;
-import io.reactivex.rxjava3.core.Flowable;
-import io.reactivex.rxjava3.functions.Consumer;
 import io.reactivex.rxjava3.schedulers.Schedulers;
 import io.reactivex.rxjava3.subjects.CompletableSubject;
 import net.sourceforge.argparse4j.ArgumentParsers;
@@ -41,14 +39,13 @@ import org.mmarini.rl.agents.BatchTrainer;
 import org.mmarini.rl.agents.KpiBinWriter;
 import org.mmarini.rl.agents.Serde;
 import org.mmarini.rl.nets.TDNetwork;
-import org.mmarini.swing.GridLayoutHelper;
+import org.mmarini.wheelly.swing.KpisPanel;
 import org.mmarini.wheelly.swing.Messages;
 import org.mmarini.yaml.Locator;
 import org.mmarini.yaml.Utils;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.api.rng.Random;
 import org.nd4j.linalg.factory.Nd4j;
-import org.nd4j.linalg.ops.transforms.Transforms;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,7 +58,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Predicate;
-import java.util.stream.Stream;
 
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
@@ -75,8 +71,6 @@ import static org.mmarini.yaml.Utils.fromFile;
 public class BatchTraining {
     private static final String BATCH_SCHEMA_YML = "https://mmarini.org/wheelly/batch-schema-0.1";
     private static final int KPIS_CAPACITY = 1000;
-    private static final double DISCOUNT = 0.977;
-    private static final double PPM_SCALE = 1e6;
     private static final Logger logger = LoggerFactory.getLogger(BatchTraining.class);
 
     static {
@@ -127,17 +121,13 @@ public class BatchTraining {
 
     protected final Namespace args;
     private final CompletableSubject completed;
-    private final JProgressBar infoBar;
-    private final JFormattedTextField deltaField;
-    private final JFormattedTextField counterField;
-    private final JFormattedTextField sensorActionGradField;
-    private final JFormattedTextField directionGradField;
-    private final JFormattedTextField speedGradField;
+    private final JTextField infoBar;
+    private final KpisPanel kpisPanel;
+    private final JProgressBar recordBar;
     private String pathFile;
     private BatchTrainer trainer;
     private boolean backUp;
     private KpiBinWriter kpiWriter;
-    private Map<String, Consumer<INDArray>> handlers;
 
     /**
      * Creates the application
@@ -147,107 +137,23 @@ public class BatchTraining {
     public BatchTraining(Namespace args) {
         this.args = requireNonNull(args);
         this.completed = CompletableSubject.create();
-        this.infoBar = new JProgressBar(JProgressBar.HORIZONTAL);
-        this.counterField = new JFormattedTextField();
-        this.deltaField = new JFormattedTextField();
-        this.sensorActionGradField = new JFormattedTextField();
-        this.speedGradField = new JFormattedTextField();
-        this.directionGradField = new JFormattedTextField();
+        this.infoBar = new JTextField();
+        this.kpisPanel = new KpisPanel();
+        this.recordBar = new JProgressBar(JProgressBar.HORIZONTAL);
         init();
     }
 
     /**
-     * Returns the action UI and kpi handlers
-     *
-     * @param key the action key
-     */
-    private Tuple2<Container, Map<String, Consumer<INDArray>>> createActionUIHandlers(String key) {
-        JFormattedTextField gradField = new JFormattedTextField();
-        gradField.setValue(0D);
-        gradField.setColumns(10);
-        gradField.setEditable(false);
-        gradField.setHorizontalAlignment(JTextField.RIGHT);
-
-        JFormattedTextField probField = new JFormattedTextField();
-        probField.setValue(0D);
-        probField.setColumns(10);
-        probField.setEditable(false);
-        probField.setHorizontalAlignment(JTextField.RIGHT);
-
-        JFormattedTextField probRatioField = new JFormattedTextField();
-        probRatioField.setValue(1D);
-        probRatioField.setColumns(10);
-        probRatioField.setEditable(false);
-        probRatioField.setHorizontalAlignment(JTextField.RIGHT);
-
-        JPanel panel = new GridLayoutHelper<>(Messages.RESOURCE_BUNDLE, new JPanel())
-                .modify("insets,4, e weight,1,1 nofill")
-                .modify("at,0,0").add("BatchTraining.gradField.label")
-                .modify("at,0,1").add(gradField)
-                .modify("at,0,2").add("BatchTraining.probField.label")
-                .modify("at,0,3").add(probField)
-                .modify("at,0,4").add("BatchTraining.probRatioField.label")
-                .modify("at,0,5").add(probRatioField)
-                .getContainer();
-        panel.setBorder(BorderFactory.createTitledBorder(key));
-        AverageValue gradAverage = new AverageValue(0, DISCOUNT);
-        AverageValue probAverage = new AverageValue(0, DISCOUNT);
-        AverageValue probRatioAverage = new AverageValue(1, DISCOUNT);
-        Map<String, Consumer<INDArray>> handlers = Map.of(
-                "netGrads." + key, grads -> gradField.setValue(maxAvg(gradAverage, grads) * PPM_SCALE),
-                "policy." + key, policy -> {
-                    try (INDArray max = policy.max(0)) {
-                        try (INDArray min = policy.min(0)) {
-                            try (INDArray ratio = max.div(min)) {
-                                probField.setValue(updateAvg(probAverage, max) * 100);
-                                probRatioField.setValue(updateAvg(probRatioAverage, ratio));
-                            }
-                        }
-                    }
-                }
-        );
-        return Tuple2.of(panel, handlers);
-    }
-
-    /**
      * Returns the content
+     * @param actions the action keys
      */
-    private Component createContent(List<String> outputs) {
-        List<Tuple2<String, Tuple2<Container, Map<String, Consumer<INDArray>>>>> actions = outputs.stream()
-                .filter(Predicate.not("critic"::equals))
-                .sorted()
-                .map(key -> Tuple2.of(key, createActionUIHandlers(key)))
-                .toList();
-        JPanel criticPanel = new GridLayoutHelper<>(Messages.RESOURCE_BUNDLE, new JPanel())
-                .modify("insets,4 e weight,1,1 nofill")
-                .modify("at,0,0").add("BatchTraining.counterField.label")
-                .modify("at,0,1").add(counterField)
-                .modify("at,0,2").add("BatchTraining.deltaField.label")
-                .modify("at,0,3").add(deltaField)
-                .getContainer();
-        criticPanel.setBorder(BorderFactory.createTitledBorder(Messages.getString("BatchTraining.criticPanel.title")));
-
-        GridLayoutHelper<JPanel> panelBuilder = new GridLayoutHelper<>(Messages.RESOURCE_BUNDLE, new JPanel())
-                .modify("insets,2 center fill weight,1,1")
-                .modify("at,0,0").add(criticPanel);
-        for (int i = 0; i < actions.size(); i++) {
-            panelBuilder.at(i + 1, 0).add(actions.get(i)._2._1);
-        }
+    private Component createContent(String... actions) {
         JPanel content = new JPanel();
         content.setLayout(new BorderLayout());
-        content.add(panelBuilder.getContainer(), BorderLayout.CENTER);
-        content.add(infoBar, BorderLayout.SOUTH);
-
-        Stream<Tuple2<String, Consumer<INDArray>>> actionHandlers = actions.stream()
-                .flatMap(t -> Tuple2.stream(t._2._2));
-
-        AverageValue delta = new AverageValue(0, DISCOUNT);
-        this.handlers = Stream.concat(
-                        actionHandlers,
-                        Stream.<Tuple2<String, Consumer<INDArray>>>of(
-                                Tuple2.of("delta", grads ->
-                                        deltaField.setValue(maxAvg(delta, grads) * PPM_SCALE))))
-                .collect(Tuple2.toMap());
+        kpisPanel.setKeys(actions);
+        content.add(kpisPanel, BorderLayout.CENTER);
+        content.add(infoBar, BorderLayout.NORTH);
+        content.add(recordBar, BorderLayout.SOUTH);
         return content;
     }
 
@@ -269,7 +175,7 @@ public class BatchTraining {
      */
     private void info(String fmt, Object... args) {
         String msg = format(fmt, args);
-        infoBar.setString(msg);
+        infoBar.setText(msg);
         logger.atInfo().log(msg);
     }
 
@@ -277,35 +183,11 @@ public class BatchTraining {
      * Initializes the application
      */
     private void init() {
-        Stream.of(
-                deltaField,
-                sensorActionGradField,
-                speedGradField,
-                counterField,
-                directionGradField
-        ).forEach(f -> f.setEditable(false));
-
-        Stream.of(
-                deltaField,
-                sensorActionGradField,
-                speedGradField,
-                counterField,
-                directionGradField
-        ).forEach(f -> {
-            f.setColumns(10);
-            f.setHorizontalAlignment(JTextField.RIGHT);
-        });
-        Stream.of(
-                deltaField,
-                sensorActionGradField,
-                speedGradField,
-                directionGradField
-        ).forEach(f -> f.setValue(0D));
-
-        counterField.setValue(0L);
-
-        infoBar.setIndeterminate(true);
-        infoBar.setStringPainted(true);
+        recordBar.setValue(0);
+        recordBar.setStringPainted(true);
+        infoBar.setHorizontalAlignment(JTextField.CENTER);
+        infoBar.setFont(infoBar.getFont().deriveFont(Font.BOLD));
+        infoBar.setEditable(false);
     }
 
     /**
@@ -335,16 +217,6 @@ public class BatchTraining {
     }
 
     /**
-     * Returns the average max value
-     *
-     * @param average the average accumulator
-     * @param values  the values
-     */
-    private double maxAvg(AverageValue average, INDArray values) {
-        return updateAvg(average, Transforms.abs(values).max(1));
-    }
-
-    /**
      * Runs the batch
      *
      * @throws Exception in case of error
@@ -352,7 +224,9 @@ public class BatchTraining {
     private void runBatch() throws Exception {
         // Prepares for training
         info("Preparing for training ...");
-        trainer.prepare(new File(this.args.getString("dataset")));
+        trainer.validate(new File(this.args.getString("dataset")));
+        recordBar.setMaximum((int) trainer.numRecords());
+        trainer.prepare();
 
         // reads kpis if activated
         String kpiPath = this.args.getString("kpis");
@@ -417,7 +291,7 @@ public class BatchTraining {
         JsonSchemas.instance().validateOrThrow(config, BATCH_SCHEMA_YML);
 
         // Creates random number generator
-        Random random = Nd4j.getRandom();
+        Random random = Nd4j.getRandomFactory().getNewRandomInstance();
         long seed = Locator.locate("seed").getNode(config).asLong(0);
         if (seed > 0) {
             random.setSeed(seed);
@@ -431,9 +305,11 @@ public class BatchTraining {
         Map<String, Float> alphas = loadAlphas();
 
         // Create the frame
+        String[] actions = outputs.stream()
+                .filter(Predicate.not("critic"::equals))
+                .toArray(String[]::new);
         JFrame frame = createFrame(Messages.getString("BatchTraining.title"),
-                new Dimension(400, 300),
-                createContent(outputs));
+                createContent(actions));
         frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 
         // Create the batch trainer
@@ -454,13 +330,12 @@ public class BatchTraining {
                 .subscribe();
         trainer.readKpis()
                 .observeOn(Schedulers.io())
-                .flatMap(kpis -> Flowable.fromIterable(kpis.entrySet()))
-                .filter(kpis -> handlers.containsKey(kpis.getKey()))
-                .doOnNext(t -> handlers.get(t.getKey()).accept(t.getValue()))
+                .doOnNext(kpisPanel::addKpis)
                 .subscribe();
         trainer.readCounter()
                 .observeOn(Schedulers.io())
-                .doOnNext(counterField::setValue)
+                .map(Number::intValue)
+                .doOnNext(recordBar::setValue)
                 .subscribe();
 
         Thread hook = new Thread(this::handleShutdown);
@@ -475,16 +350,4 @@ public class BatchTraining {
                 .subscribe();
     }
 
-    /**
-     * Returns the updated average
-     *
-     * @param average the average value
-     * @param values  the values
-     */
-    private double updateAvg(AverageValue average, INDArray values) {
-        for (long i = 0; i < values.size(0); i++) {
-            average.add(values.getDouble(i));
-        }
-        return average.getValue();
-    }
 }